Refactor: 리그 시즌 훅 생성
