+++ src/components/match/match-record.tsx
@@ -4,7 +4,7 @@ import Box from '@/components/common/box';
 import MatchRecordTab from '@/components/match/match-record-tab';
 import MatchRecordCategory from '@/components/match/match-record-category';
 import MomVote from '@/components/match/mom-vote';
-import PlayerRank from '@/components/player-rank/player-rank';
+import PlayerRankTable from '@/components/player-rank/player-rank-table';
 
 import Chelsea from '@/assets/Chelsea.png';
 import MatchSelectTeamButton from '@/components/match/match-select-team-button';
@@ -52,7 +52,7 @@ export default function MatchRecord({ home, away }: MatchRecordProps) {
           </li>
         </ul>
         <TableContainer isMatch={true}>
-          <PlayerRank />
+          <PlayerRankTable />
         </TableContainer>
       </Box>
     </div>
