+++ src/utils/leagueNameMapper.ts
@@ -1,0 +1,4 @@
+export const leagueNameMapper = {
+  premier: '프리미어리그',
+  laliga: '라리가',
+};
